pyproject.toml
[project]
name = "3gppschedule"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-docx>=1.1.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.2.1",
]
